Circular dependency fix
